tg-606: refinement in script

+++ .swarm/save-db.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 
-_now=$(date +"%Y%m%dT%H%M")
-mv /tmp/exported_regions.sql /tmp/exported_regions_$_now.sql
+if [ -f /tmp/exported_regions.sql ]; then
+	_now=$(date +"%Y%m%dT%H%M")
+	mv /tmp/exported_regions.sql /tmp/exported_regions_$_now.sql
+fi
 
 _db_docker=`docker ps -q -f "name=panoswarm_database"`
 
@@ -10,4 +12,4 @@ echo EXPORTING DATA FROM CONTAINTER $_db_docker
 docker exec -it $_db_docker psql -U panorama -c 'drop table if exists region_export'
 docker exec -it $_db_docker psql -U panorama -c 'create table region_export as select region_type, left_top_x, left_top_y, right_top_x, right_top_y, left_bottom_x, left_bottom_y, right_bottom_x, right_bottom_y, detected_by, pano_id from panoramas_region'
 docker exec -it $_db_docker pg_dump -Fc -t region_export* -d panorama -U panorama -f /tmp/exported_regions.sql
-docker cp $_db_docker:/tmp/exported_regions.sql /tmp/
+docker cp $_db_docker:/tmp/exported_regions.sql /tmp/exported_regions.sql
